add dynamic loadable plugins
The advantage of dynamic loadable plugins is that all their libraries
are no longer needed when not using them and when using one only its
own libraries need to be loaded. As a side effect that allows
distributions to make packages that support all those methods without
the package needing to depend on all the libraries (a simple Suggests
instead of an Depend in Debian), so people can install lprng without
needing to install kerberos libraries and the libraries kerberos
libraries depend upon.

The current implementation is quite low-tech and not very nice but
functional:

1) It's not using libtool but just tries some compiler flags and if
they do not work falls back to do without plugins. As there is a
easy fallback (just compile in as done previously), having all the
horrors of libtool (config.guess/sub files, all those libtool scripts,
checking for fortran in configure) are avoided, as none of the features
are needed.

2) The generated plugins just use all functions and global variables
from the binary itself. This has the disadvantage that the binary has
to be compiled with -export-dynamic and the plugins cannot be compiled
when -Wl,-z,defs is in the LDFLAGS and that it looks a bit dirty.
But changing that means to rewrite most of LPRngs file handling
(which is a good idea anyway, but definitly nothing possible
 middle- or even short-term).

The patch in its current form does not enable plugins by default but
only if compiled with --enable-plugins and should not change anything
when that is not given.

If enabled, it should not change behaviour (if there are no bugs
hidden), except in one case: It currently does no
longer output a list of all authentication possibilites. That could
be readded, but would most likely mean to load all authentication
plugins to get that information. perhaps just listing in which
directly to look would be a more sensible solution.