improve checking for kerberos 5 libraries
Check at configure time that krb5_encrypt_size is available to make
sure the correct kerberos crypto library is loaded.

Check first for k5crypto, as that is a more special name so more
likely to be the used library.

If none of the guesses to get the libraries right works, try calling
krb5-config. (That outputs things like -L/usr/lib, so not use it unless
none of the guesses works).