drop --with-ssl, use autoconf macros to look for libssl
configure looked for libssl.so and libssl.a files in a list of
hard-coded directories or could be given a path to look for them
(but not a list).

This breaks if libraries are in other directories (like the are
in some distributions due to "multiarch" now).

So simply use all the manual work and use autoconf macros to
look for headers and libraries, as they do not look for files
but check if the compiler can find those files.

Also drop --with-ssl, people having libraries in locations their
linker cannot find them can still use CPPFLAGS and LDFLAGS
to tell the compiler where those are.

Additionally configure fails now if --enable-ssl is given
explicitly and it cannot be enabled.
(It was silently disabled before).